* Adjusted store styling
